Move ViewHomework Test class to correct package.
